flask-project/requirements.txt
flask==1.1.2
python-dotenv
requests
gunicorn